allows user to set the IP address of their robot
IP Address of robot was previously set, as a final, to an arbitrary IP
address. This change allows a user to input the IP address of their
robot when the program starts. Especially useful if IP Address of robot
is constantly changing. Program shuts down if user does not enter an IP
Address.